training/sft_dpo/utils/requirements_local.txt
accelerate==1.3.0
datasets==3.2.0
google-genai==1.12.1
langdetect==1.0.9
numpy==1.26.4
openai==1.93.3
torch==2.5.1
tqdm==4.67.1
transformers==4.48.3
trl==0.14.0
vllm==0.7.0
